service基类增加新增和修改方法

+++ AhDai.Db/Models/BaseModel.cs
@@ -5,7 +5,7 @@ namespace AhDai.Db.Models;
 /// <summary>
 /// BaseModel
 /// </summary>
-public class BaseModel : BaseModel<int>
+public class BaseModel : BaseModel<long>
 {
 }
