+++ AhDai.Db/Models/DictDatum.cs
@@ -8,7 +8,7 @@ public partial class DictDatum : BaseModel
 	/// <summary>
 	/// 字典Id
 	/// </summary>
-	public int DictId { get; set; }
+	public long DictId { get; set; }
 	/// <summary>
 	/// 编码
 	/// </summary>
